Fixing a race condition
When enabling debug mode on the iOS simulator, we attempt to modify a plist file backing an NSUserDefaults object. This plist file is written to by the app at various times, and this would sometimes overwrite our change.

With this fix, we start the app to ensure the file is created, then we stop the app to avoid a race condition. We need to wait a moment to ensure that the app is completely closed, then we write to the file and restart the app. Simply relaunching the app is not sufficient, we need to re-install the app or it does not pick up the changes.