Fixing an issue with 'this' in ios/deviceRunner Fixing iOS launch when device is in debug mode
When in debug mode the device does not produce any console output. We considered a launch successful when we saw the first message produced by the device, which would mean that the device launch would time out. With this change we consider the device started as soon as we send the final of the 3 messages to launch an app.

Updating test to account for new startup protocol