Defer output channel conversion
Instead of creating an output channel once our extension is loaded, with this change we defer it to when it is required. When the output channel is created it triggers the output pane to come into view, so by defering it we avoid cluttering the screen until there is something worth looking at.