Updating openFileAtLocation to expect correct arguments
The "atom" scripts that we used to open files when clicking on stack traces were not correctly set as executable. To fix this, I now explicitly run them via node. In addition, react-native may or may not pass a workspace variable, so we now check for this to handle it apropriately.